+++ core/src/main/java/com/graphhopper/routing/util/countryrules/europe/HungaryCountryRule.java
@@ -18,7 +18,7 @@
 package com.graphhopper.routing.util.countryrules.europe;
 
 import com.graphhopper.reader.ReaderWay;
-import com.graphhopper.routing.ev.RoadAccess;
+import com.graphhopper.routing.ev.CarRoadAccess;
 import com.graphhopper.routing.ev.RoadClass;
 import com.graphhopper.routing.ev.Toll;
 import com.graphhopper.routing.util.TransportationMode;
@@ -32,22 +32,22 @@ import com.graphhopper.routing.util.countryrules.CountryRule;
 public class HungaryCountryRule implements CountryRule {
 
     @Override
-    public RoadAccess getAccess(ReaderWay readerWay, TransportationMode transportationMode, RoadAccess currentRoadAccess) {
+    public CarRoadAccess getAccess(ReaderWay readerWay, TransportationMode transportationMode, CarRoadAccess currentCarRoadAccess) {
         // Pedestrian traffic and bicycles are not restricted
 		if (transportationMode == TransportationMode.FOOT || transportationMode == TransportationMode.BIKE) {
-            return currentRoadAccess;
+            return currentCarRoadAccess;
         }
 
-        if (currentRoadAccess != RoadAccess.YES) {
-            return currentRoadAccess;
+        if (currentCarRoadAccess != CarRoadAccess.YES && currentCarRoadAccess != CarRoadAccess.MISSING) {
+            return currentCarRoadAccess;
         }
 
         RoadClass roadClass = RoadClass.find(readerWay.getTag("highway", ""));
         if (roadClass == RoadClass.LIVING_STREET) {
-            return RoadAccess.DESTINATION;
+            return CarRoadAccess.DESTINATION;
         }
 
-        return RoadAccess.YES;
+        return CarRoadAccess.YES;
     }
 
     @Override