+++ web/src/test/java/com/graphhopper/application/resources/RouteResourceCustomModelTest.java
@@ -65,7 +65,7 @@ public class RouteResourceCustomModelTest {
                 putObject("graph.location", DIR).
                 putObject("custom_areas.directory", "./src/test/resources/com/graphhopper/application/resources/areas").
                 putObject("import.osm.ignored_highways", "").
-                putObject("graph.encoded_values", "car_access, car_average_speed, road_access, " +
+                putObject("graph.encoded_values", "car_access, car_average_speed, car_road_access, " +
                         "bike_access, bike_priority, bike_average_speed, bike_road_access, " +
                         "foot_access, foot_priority, foot_average_speed, foot_road_access, " +
                         "max_height, max_weight, max_width, hazmat, toll, surface, track_type, hgv, " +